% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{create_mds_or_pca_plot}
\alias{create_mds_or_pca_plot}
\title{Creates a MDS or PCA plot by passing GLimma::glMDSPlot output to ggplot2}
\usage{
create_mds_or_pca_plot(
  loc_exp_lcpm,
  top = 500,
  dimensions = c(1, 2),
  gene.selection = "pairwise",
  factor_one = exp_DGElist$samples$treatment,
  factor_two = exp_DGElist$samples$RNAi,
  Title = "MDS plot using avg(logFC) as the distance",
  geom_size = 2.5,
  shape_legend_color = "grey"
)
}
\arguments{
\item{loc_exp_lcpm}{Log transformed count matrix.}

\item{top}{Top number of genes to be analysed e.g. top 500 genes}

\item{dimensions}{PC dimensions to be plotted.}

\item{gene.selection}{"pairwise" for MDS. "common" for PCA.}

\item{factor_one}{Column of DGElist$samples that contains first factor to be labelled by color e.g. treatment}

\item{factor_two}{Column of DGElist$samples that contains second factor to be labelled by shape e.g. RNAi}

\item{shape_legend_color}{}
}
\value{
plot
}
\description{
Creates a MDS or PCA plot by passing GLimma::glMDSPlot output to ggplot2
}
