% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_helper_functions.R
\name{creating_gene_sets}
\alias{creating_gene_sets}
\title{Extract gene set character vectors from list of character vectors.}
\usage{
creating_gene_sets(
  list_of_tbls,
  gene_id = "entrez_gene",
  gene_set_name = "gs_name"
)
}
\arguments{
\item{list_of_tbls}{e.g. from dplyr group_split function.}

\item{gene_set_name}{}
}
\value{
list of gene set character vectors.
}
\description{
Extract gene set character vectors from list of character vectors.
}
