% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_diff_expressed_genes.R
\name{differential_expressed_genes_per_contrast_char}
\alias{differential_expressed_genes_per_contrast_char}
\title{Wrapper function for Limma TopTable function. Returns vector of genes ids.}
\usage{
differential_expressed_genes_per_contrast_char(
  p_val_cutoff,
  LFC_cutoff,
  gene_id_of_choice = "flybase_ids",
  direction = "Both",
  ...
)
}
\arguments{
\item{gene_id_of_choice}{Can be flybase id, symbol etc.}

\item{direction}{up, down or both}

\item{...}{Passing the dots to TopTable function.}
}
\value{

}
\description{
Wrapper function for Limma TopTable function. Returns vector of genes ids.
}
