% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_diff_expressed_genes.R
\name{differential_expressed_genes_per_contrast_df}
\alias{differential_expressed_genes_per_contrast_df}
\title{Wrapper function for Limma TopTable function. Returns dataframe.}
\usage{
differential_expressed_genes_per_contrast_df(p_val_cutoff, LFC_cutoff, ...)
}
\arguments{
\item{...}{Passing the dots to TopTable function.}
}
\value{
Data Frame of differentially expressed genes from toptable.
}
\description{
Wrapper function for Limma TopTable function. Returns dataframe.
}
