% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGElist_edit_functions.R
\name{edit_DGElist_genes}
\alias{edit_DGElist_genes}
\title{Add extra gene info to DGElist object.}
\usage{
edit_DGElist_genes(
  exp_DGElist,
  org.Dm.eg.db,
  desired_columns = c("FLYBASE", "SYMBOL", "ONTOLOGY", "PATH", "ENTREZID"),
  extra_gene_info = NA,
  extra_gene_type_name = NA,
  return_genes_info_only = TRUE
)
}
\arguments{
\item{exp_DGElist}{DGElist object from edgeR package}

\item{return_genes_info_only}{}
}
\value{

}
\description{
Add extra gene info to DGElist object.
}
