% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_helper_functions.R
\name{multiple_constrast_camera}
\alias{multiple_constrast_camera}
\title{Multiple contrast wrapper for limma camera function.}
\usage{
multiple_constrast_camera(contr_matrix, ...)
}
\arguments{
\item{...}{Passing dots to limma camera function.}
}
\value{
List of camera tibbles. Can later bind in one tibble using dplyr::bind_rows(.id = "contrast").
}
\description{
Multiple contrast wrapper for limma camera function.
}
