% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_helper_functions.R
\name{multiple_constrast_mroast}
\alias{multiple_constrast_mroast}
\title{Multiple contrast wrapper for mroast camera function.}
\usage{
multiple_constrast_mroast(contr_matrix, ...)
}
\arguments{
\item{...}{Passing dots to limma mroast function.}
}
\value{
List of mroast tibbles. Can later bind in one tibble using dplyr::bind_rows(.id = "contrast")
}
\description{
Multiple contrast wrapper for mroast camera function.
}
