% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_helper_functions.R
\name{multiple_constrast_roast}
\alias{multiple_constrast_roast}
\title{Multiple contrast roast wrapper function - not mroast}
\usage{
multiple_constrast_roast(v, gene_set, design, contr_matrix, nrot = 9999)
}
\arguments{
\item{v}{voomed DGElist}

\item{gene_set}{character vector or list of character vectors containing gene sets. Ids must match v e.g. Flybase Ids.}

\item{design}{design matrix}

\item{contr_matrix}{contrast matrix}

\item{nrot}{number of rotations to pass to roast function.}
}
\value{
df
}
\description{
Multiple contrast roast wrapper function - not mroast
}
