% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_diff_expressed_genes.R
\name{multiple_constrast_top_genes}
\alias{multiple_constrast_top_genes}
\title{Multiple contrast wrapper to identify differentially expressed genes}
\usage{
multiple_constrast_top_genes(
  constrast_matrix,
  target_function,
  annotation_for_list_names = "",
  ...
)
}
\arguments{
\item{target_function}{"differential_expressed_genes_per_contrast_char" or "differential_expressed_genes_per_contrast_df"}

\item{...}{Pass the dots to the target_function. This then passes to the TopTable function.}
}
\value{
list of character vectors or list of dataframes.
}
\description{
Multiple contrast wrapper to identify differentially expressed genes
}
