% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{multiple_reactomePA_save_dotplot_and_csv}
\alias{multiple_reactomePA_save_dotplot_and_csv}
\title{Save dotplots and csvs from reactome pathway analysis}
\usage{
multiple_reactomePA_save_dotplot_and_csv(
  flybase_multiple_gene_list,
  entrez_ids,
  start_of_file_name,
  save_directory,
  number_cat_dotplot = 15,
  font_size_dotpot = 8,
  height_dotplot = 10,
  width_dotplot = 15,
  ...
)
}
\arguments{
\item{flybase_multiple_gene_list}{Gene list of flybase ids}

\item{entrez_ids}{Column of DGElist$genes that contains entrez id info.}

\item{...}{}
}
\value{

}
\description{
Save dotplots and csvs from reactome pathway analysis
}
