% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otto_venn_diagram.R
\name{otto_venn_diagram}
\alias{otto_venn_diagram}
\title{Creates custom venn diagram based on limma package Venn diagram function.}
\usage{
otto_venn_diagram(
  object,
  include = "both",
  names = NULL,
  mar = rep(1, 4),
  cex = c(1.5, 1, 0.7),
  lwd = 1,
  circle.col = NULL,
  counts.col = NULL,
  show.include = NULL,
  col_alpha = 0.2,
  ...
)
}
\arguments{
\item{object}{from decide tests}

\item{...}{passed to plot function}
}
\value{

}
\description{
Creates custom venn diagram based on limma package Venn diagram function.
}
