% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{pheatmap_wrapper_function}
\alias{pheatmap_wrapper_function}
\title{Creates heatmap using pheatmap function. Can have two rows of labels.}
\usage{
pheatmap_wrapper_function(
  sample_df,
  x_lcpm,
  gene_set_char,
  filt_1_quo = RNAi,
  filt_1_vec = c("LacZRNAi"),
  filt_2_quo = treatment,
  filt_2_vec = NA,
  var_to_groupby = group,
  ...
)
}
\arguments{
\item{sample_df}{df from DGElist$samples}

\item{x_lcpm}{Count matrix. Usually log transformed.}

\item{gene_set_char}{Gene set as a vector Identifies need to match rownames of loc_exp_lcpm e.g. flybase_ids}

\item{filt_1_quo}{Column name in sample_df. Quosure. Do not quote.}

\item{filt_1_vec}{Vector containing elements to select from filt_1 column of sample_df}

\item{filt_2_quo}{Column name in sample_df. Quosure. Do not quote.}

\item{filt_2_vec}{Vector containing elements to select from filt_2 column of sample_df}

\item{var_to_groupby}{Normally DGElist$samples$group. Normally a Factor}

\item{...}{Passing the dots to pheatmap function.}
}
\value{

}
\description{
Creates heatmap using pheatmap function. Can have two rows of labels.
}
