% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_in_csvs.R
\name{read_in_csvs_as_list_of_char}
\alias{read_in_csvs_as_list_of_char}
\title{Reading in csvs as list of character vectors.}
\usage{
read_in_csvs_as_list_of_char(path_glob_str)
}
\arguments{
\item{path_glob_str}{Glob to folder of csv files to be read.}
}
\value{
List of character vectors e.g. Gene sets
}
\description{
Reading in csvs as list of character vectors.
}
