% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_stuff.R
\name{saving_html_datatables_from_list_of_tbls}
\alias{saving_html_datatables_from_list_of_tbls}
\title{Saving list of dataframes to html datatables.}
\usage{
saving_html_datatables_from_list_of_tbls(
  list,
  additional_annotation_string,
  directory,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Saving list of dataframes to html datatables.
}
