% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{scale_extract_gene_set_and_tidy_lcpm_data}
\alias{scale_extract_gene_set_and_tidy_lcpm_data}
\title{Scale and tidy data prior to ggplot2 plot e.g. violin plot.}
\usage{
scale_extract_gene_set_and_tidy_lcpm_data(
  loc_exp_lcpm,
  groups,
  RNAi,
  treatment,
  sam_ids,
  gene_set_charV
)
}
\arguments{
\item{loc_exp_lcpm}{Count matrix. Usually log transformed.}

\item{groups}{Column of DGElist$samples that contains group info.}

\item{RNAi}{Column of DGElist$samples that contains RNAi info.}

\item{treatment}{Column of DGElist$samples that contains treatment info.}

\item{sam_ids}{Column of DGElist$samples that contains sam_ids info.}

\item{gene_set_charV}{Gene set as a vector. Identifies need to match rownames of loc_exp_lcpm e.g. flybase_ids}
}
\value{
tidy df
}
\description{
Scale and tidy data prior to ggplot2 plot e.g. violin plot.
}
