% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{scale_extract_gene_set_and_tidy_lcpm_data_simple}
\alias{scale_extract_gene_set_and_tidy_lcpm_data_simple}
\title{Scale and tidy data prior to ggplot2 plot e.g. violin plot. Only using group info.}
\usage{
scale_extract_gene_set_and_tidy_lcpm_data_simple(
  loc_exp_lcpm,
  groups,
  sam_ids,
  gene_set_charV
)
}
\arguments{
\item{loc_exp_lcpm}{Log transformed count matrix.}

\item{groups}{Column of DGElist$samples that contains group info.}

\item{sam_ids}{Column of DGElist$samples that contains sam_ids info.}

\item{gene_set_charV}{Gene set as a vector. Identifies need to match rownames of loc_exp_lcpm e.g. flybase_ids}
}
\value{
tidy_df
}
\description{
Scale and tidy data prior to ggplot2 plot e.g. violin plot. Only using group info.
}
