% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_plotting.R
\name{violinplot_from_tidy_gene_set_data}
\alias{violinplot_from_tidy_gene_set_data}
\title{Violin plot using tidy data}
\usage{
violinplot_from_tidy_gene_set_data(tidy_gene_set_data)
}
\arguments{
\item{tidy_gene_set_data}{}
}
\value{

}
\description{
Violin plot using tidy data
}
