% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_stuff.R
\name{write_list_2_csv_directory}
\alias{write_list_2_csv_directory}
\title{Saving list of dataframes to csvs.}
\usage{
write_list_2_csv_directory(
  list,
  additional_annotation_string = "",
  directory,
  ...
)
}
\arguments{
\item{...}{Passing the dots write.table function.}
}
\value{

}
\description{
Saving list of dataframes to csvs.
}
